+++ wavefunction/__init__.py
@@ -1,8 +1,10 @@
 from .wavefunction import Wavefunction
 from .wavefunction import SpectralMode
+from .wavefunction import Basis
 
 
 __all__ = [
     "Wavefunction",
     "SpectralMode",
+    "Basis"
 ]